auth sameinað með wishlist, AppNavGraph helsta breytingin
