+++ app/src/main/java/is/hi/present/navigation/Routes.kt
@@ -1,6 +1,7 @@
 package `is`.hi.present.navigation
 
 object Routes {
+    const val AUTH = "auth"
     const val WISHLISTS = "wishlists"
     const val CREATE_WISHLIST = "create_wishlist"
     const val WISHLIST_DETAIL = "wishlist/{wishlistId}"
