--- app/src/main/java/is/hi/present/ui/auth/LoginScreen.kt
@@ -1,4 +1,0 @@
-package `is`.hi.present.ui.auth
-
-class LoginScreen {
-}
